+++ src/signature.rs
@@ -30,6 +30,36 @@ impl Signature {
             }),
         )
     }
+
+    /// Returns true if the inner point is free of an $h$-torsion component, and
+    /// so it exists within the $q$-order subgroup $\mathbb{G}_2$. This
+    /// should always return true unless an "unchecked" API was used.
+    pub fn is_torsion_free(&self) -> bool {
+        self.0.is_torsion_free().into()
+    }
+
+    /// Returns true if the inner point is on the curve. This should always
+    /// return true unless an "unchecked" API was used.
+    pub fn is_on_curve(&self) -> bool {
+        self.0.is_on_curve().into()
+    }
+
+    /// Returns true if the inner point is the identity (the point at infinity).
+    pub fn is_identity(&self) -> bool {
+        self.0.is_identity().into()
+    }
+
+    /// Returns true if the inner point is valid according to certain criteria.
+    ///
+    /// A [`Signature`] is considered valid if its inner point meets the
+    /// following conditions:
+    /// 1. It is free of an $h$-torsion component and exists within the
+    ///    $q$-order subgroup $\mathbb{G}_2$.
+    /// 2. It is on the curve.
+    /// 3. It is not the identity.
+    pub fn is_valid(&self) -> bool {
+        self.is_torsion_free() && self.is_on_curve() && !self.is_identity()
+    }
 }
 
 impl Serializable<48> for Signature {