+++ CHANGELOG.md
@@ -12,6 +12,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Change the implementation for hashing a slice of bytes into a BlsScalar to `BlsScalar::hash_to_scalar` [#3]
+- Check validity of `PublicKey` and `Signature` points in signature verification [#7]
+
+### Added
+
+- Add `is_valid` check for `PublicKey` [#7]
+- Add `Error::InvalidPoint` variant for invalid `PublicKey` and `Signature` points [#7]
 
 ## [0.1.0] - 2024-01-08
 
@@ -20,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add initial commit, this package continues the development of [dusk-bls12_381-sign](https://github.com/dusk-network/bls12_381-sign/) at version `0.6.0` under the new name: `bls12_381-bls` and without the go related code.
 
 <!-- ISSUES -->
+[#7]: https://github.com/dusk-network/bls12_381-bls/issues/7
 [#3]: https://github.com/dusk-network/bls12_381-bls/issues/3
 
 <!-- VERSIONS -->